requirements.txt
akismet==1.2.1
anyio==3.6.1
asgiref==3.5.0
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.21.30
botocore==1.24.30
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
cryptography==37.0.4
decorator==5.1.1
Django==3.2.12
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-datatables-too==0.1.4
django-extra-views==0.14.0
django-js-asset==1.2.2
django-multi-form-view==2.0.1
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
jmespath==1.0.0
outcome==1.2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
querystring-parser==1.2.4
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.5.2
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.8
validators==0.19.0
wsproto==1.1.0
youtube-search-python==1.6.6
